Add hmmm package: #PapersThatMakeYouGoHmmm.
This package extracts the past week's worth of papers from a given
arXiv subtopic.  It provides a method on Papr to return the ArxivURL,
as well as the URLs needed to bookmark them on ScienceWise and Reddit.

Bibsonomy and Mendeley both require unique tokens per URL to bookmark,
so they were excluded from this to reduce load on the arXiv servers.